Removed get_token fn and changed return type.

+++ src/servrian/login.h
@@ -14,7 +14,7 @@
 #include "aux.h"
 
 /* Receives client's auth parameters and checks */
-int authorization(unsigned char *auth_code); 
+short authorization(unsigned char *auth_code); 
 
 #endif
